+++ tests/test_sparse.py
@@ -1,7 +1,13 @@
+import os
+
 import numpy as np
 import pytest
 
 from xugrid.core import sparse
+
+
+def numba_enabled() -> bool:
+    return os.environ.get("NUMBA_DISABLE_JIT") != "1"
 
 
 @pytest.fixture(scope="function")
@@ -37,11 +43,39 @@ def test_weight_matrix_csr(csr_matrix):
     assert csr_matrix.nnz == 10
 
 
+@pytest.mark.skipif(
+    numba_enabled(),
+    reason="Numba cannot convert native range_state_int64 to Python object.",
+)
 def test_nzrange(csr_matrix):
+    # These functions work fine if called inside of other numba functions when
+    # numba is enabled.
     i = sparse.nzrange(csr_matrix, 0)
     assert np.array_equal(i, range(0, 2))
     i = sparse.nzrange(csr_matrix, 1)
     assert np.array_equal(i, range(2, 4))
+
+
+@pytest.mark.skipif(
+    numba_enabled(),
+    reason="Function returns a slice object; python and no-python slices don't mix.",
+)
+def test_row_slice(csr_matrix):
+    # These functions work fine if called inside of other numba functions when
+    # numba is enabled.
+    assert sparse.row_slice(csr_matrix, 0) == slice(0, 2, None)
+
+
+@pytest.mark.skipif(
+    numba_enabled(),
+    reason="Function returns a zip object; python and no-python zips don't mix.",
+)
+def test_columns_and_values(csr_matrix):
+    # These functions work fine if called inside of other numba functions when
+    # numba is enabled.
+    zipped = sparse.columns_and_values(csr_matrix, sparse.row_slice(csr_matrix, 0))
+    result = list(zipped)
+    assert result == [(0, 0.5), (1, 0.5)]
 
 
 def test_coo_to_csr(coo_matrix):